refactor(tiling): split TilingSettings resolution helpers
TilingSettings crossed the 350-line ceiling with the space.icon
group; move the per-space resolution and interactive-resize
write helpers into TilingSettings+Resolution.swift. No behavior
change.
